+++ src/main/java/com/book/store/Service/CustomerService.java
@@ -16,21 +16,17 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 public class CustomerService {
-
     private final CustomerRepository customerRepository;
     private final CustomerMapper customerMapper;
-
     // Create customer from API DTO
     public Customer createCustomer(Customer customer) {
         return customerRepository.save(customer);
     }
-
     // Delete customer
     @Transactional
-    public void deleteCustomer(Long id) {
+    public void deleteCustomer(Integer id) {
         customerRepository.deleteById(id);
     }
-
     // List all customers
     public List<CustomerApiDto> getAllCustomers(int page, int size) {
         Pageable pageable = PageRequest.of(page, size);
@@ -38,9 +34,8 @@ public class CustomerService {
                 .map(customerMapper::toDTO)
                 .collect(Collectors.toList());
     }
-
     // Find by ID
-    public CustomerApiDto findCustomerById(Long id) {
+    public CustomerApiDto findCustomerById(Integer id) {
         Customer customer = customerRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Customer not found"));
         return customerMapper.toDTO(customer);
